+++ libs/migrations/migrations/00006_create_orders_books_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS orders_books (
     order_id INTEGER NOT NULL,
     book_id INTEGER NOT NULL,
     amount INTEGER NOT NULL,
-    price INTEGER NOT NULL,
+    price NUMERIC NOT NULL,
 
     PRIMARY KEY (order_id, book_id),
     FOREIGN KEY (order_id) REFERENCES orders(id) ON DELETE CASCADE,
